More Error changes
Moved formatting of the error message to HumanConsumableErrors themselves and outputting of error messages to GrammarOption so that errors in STLR are reported too.
